+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Changes
 ## ng-vcl
 - (BREAKING) requires Angular ^5
+- select/dropdown/metalist:
+  - (BREAKING) Removed selected property from items/options. Use value property on vcl-metalist, vcl-dropdown or vcl-select instead.
 - password-input:
   - (BREAKING) Must be used as container with vcl-input
 - link: